+++ packages/website/src/pages/layout/BrowserPreview.tsx
@@ -1,104 +1,59 @@
 import * as React from 'react';
-import {
-    BrowserPreview,
-    BrowserPreviewEmpty,
-    BrowserPreviewError,
-    Section,
-    SplitLayout,
-    Svg,
-} from '@coveord/plasma-react';
 
-import PlasmaComponent from '../../building-blocs/PlasmaComponent';
+import {PageLayout} from '../../building-blocs/PageLayout';
 
-// start-print
-export const BrowserPreviewExamples: React.FunctionComponent = () => (
-    <PlasmaComponent
+const code = `
+    import * as React from 'react';
+    import {BrowserPreview} from '@coveord/plasma-react';
+
+    export default () => (
+        <BrowserPreview title={'Custom title! Custom title!'}>   
+            <h4>Hello World</h4>
+            <p>Here's the description</p>
+        </BrowserPreview>
+    );
+`;
+
+const withError = `
+    import * as React from 'react';
+    import {BrowserPreview, BrowserPreviewError} from '@coveord/plasma-react';
+
+    export default () => (
+        <BrowserPreview>
+            <BrowserPreviewError errorMessage="The Query expression within the Reference Cases field is not valid." />
+        </BrowserPreview>
+    );
+`;
+
+const emptyState = `
+    import * as React from 'react';
+    import {BrowserPreview, BrowserPreviewEmpty} from '@coveord/plasma-react';
+
+    export default () => (
+        <BrowserPreview>
+            <BrowserPreviewEmpty onClick={() => alert('Clicked!')}>
+                <span>
+                    Select a <span className="bolder">Query Pipeline</span>
+                </span>
+                <span>to see the preview</span>
+            </BrowserPreviewEmpty>
+        </BrowserPreview>
+    );
+`;
+
+export const BrowserPreviewExamples = () => (
+    <PageLayout
         id="BrowserPreview"
-        title="Browser Preview"
-        usage="A browser preview shows the result of configuration changes in a simplified representation of a browser interface."
-        withSource
-    >
-        <Section>
-            <Section level={2} title="Browser Preview container">
-                <SplitLayout
-                    leftChildren={<div />}
-                    rightChildren={
-                        <div className="p3">
-                            <BrowserPreview />
-                        </div>
-                    }
-                />
-            </Section>
-            <Section level={2} title="Browser Preview with custom title">
-                <SplitLayout
-                    leftChildren={<div />}
-                    rightChildren={
-                        <div className="p3">
-                            <BrowserPreview title={'Custom title! Custom title!'} />
-                        </div>
-                    }
-                />
-            </Section>
-            <Section level={2} title="Browser Preview with custom content">
-                <SplitLayout
-                    leftChildren={<div />}
-                    rightChildren={
-                        <div className="p3">
-                            <BrowserPreview>
-                                <h4>Hello World</h4>
-                                <p>Here's the description</p>
-                            </BrowserPreview>
-                        </div>
-                    }
-                />
-            </Section>
-            <Section level={2} title="Browser Preview with error message">
-                <SplitLayout
-                    leftChildren={<div />}
-                    rightChildren={
-                        <div className="p3">
-                            <BrowserPreview>
-                                <BrowserPreviewError errorMessage="The Query expression within the Reference Cases field is not valid." />
-                            </BrowserPreview>
-                        </div>
-                    }
-                />
-            </Section>
-            <Section level={2} title="Browser Preview with empty state">
-                <SplitLayout
-                    leftChildren={<div />}
-                    rightChildren={
-                        <div className="p3">
-                            <BrowserPreview>
-                                <BrowserPreviewEmpty onClick={() => alert('Clicked!')}>
-                                    <span>
-                                        Select a <span className="bolder">Query Pipeline</span>
-                                    </span>
-                                    <span>to see the preview</span>
-                                </BrowserPreviewEmpty>
-                            </BrowserPreview>
-                        </div>
-                    }
-                />
-            </Section>
-            <Section level={2} title="Browser Preview with empty state and custom image">
-                <SplitLayout
-                    leftChildren={<div />}
-                    rightChildren={
-                        <div className="p3">
-                            <BrowserPreview>
-                                <BrowserPreviewEmpty
-                                    onClick={() => alert('Wow!')}
-                                    image={<Svg svgName="arrowTopSlim" className="block" />}
-                                >
-                                    <span>Look at this browser-like header!</span>
-                                </BrowserPreviewEmpty>
-                            </BrowserPreview>
-                        </div>
-                    }
-                />
-            </Section>
-        </Section>
-    </PlasmaComponent>
+        title="Browser preview"
+        section="Layout"
+        thumbnail="placeholder"
+        description="A browser preview shows the result of configuration changes in a simplified representation of a browser interface."
+        componentSourcePath="/browserPreview/BrowserPreview.tsx"
+        code={code}
+        examples={{
+            withError: {code: withError, title: 'With error'},
+            emptyState: {code: emptyState, title: 'Empty state'},
+        }}
+    />
 );
 export default BrowserPreviewExamples;